requirements.txt
alembic==1.13.1
asgiref==3.8.1
asyncpg==0.29.0
attrs==23.2.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
greenlet==3.0.3
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
setuptools==70.0.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
